I have done much things. Orz!

+++ plugit/core/Transaction.js
@@ -1,78 +1,97 @@
 'use strict';
 
-const TransactionModel = require('../plugitModels/TransactionModel');
+const mongoose = require('mongoose');
 const ObjectId = require('mongodb').ObjectId;
 const path = require('path');
-const assert = require('assert');
+const PlugitError = require('../utils/PlugitError');
 
 class Transaction {
 
-  constructor(id) {
-    this._id = id;
+  constructor(models = {}, components) {
+    const model = models.Transaction;
+    if (!(model.base instanceof mongoose.constructor)) throw new PlugitError('model must be an instance of mongoose model');
+    if (typeof components !== 'object') throw new PlugitError('components should be an object');
+    this._model = model;
+    this._models = models;
+    this._components = components;
   }
 
   get id() {
     return this._id;
   }
 
-  * info() {
-    return yield TransactionModel.findById(this._id);
+  get model() {
+    return this._model;
   }
 
-  * _create() {
+  get models() {
+    return this._models;
+  }
+
+  get components() {
+    return this._components;
+  }
+
+  * info() {
+    return yield this.model.findById(this.id);
+  }
+
+  * create() {
     if (!(yield this.info())) {
-      this._id = (yield TransactionModel().save()).id;
+      this._id = (yield this.model().save()).id;
     }
     return this;
   }
 
   * pend() {
     const {state} = yield this.info();
-    assert(state == 'init', 'Transaction state is not init');
-    yield TransactionModel.findByIdAndUpdate(this._id, { $set: { state: 'pendding' } });
+    if (state !== 'init') throw new PlugitError('Transaction state is not init');
+    yield this.model.findByIdAndUpdate(this._id, { $set: { state: 'pendding' } });
   }
 
-  * pushAction({component, operation}, ...params) {
-    assert(operation, 'Action should have an operation!');
+  * run({component, operation}, ...params) {
+    const Component = require('../base/Component');
+    if (!(component instanceof Component)) throw new PlugitError('component must be an instance of Component or class extends Component');
+    if (typeof operation !== 'string' || !operation) throw new PlugitError('Action should have an string operation');
     //Check the transaction state;
     const {state} = yield this.info();
-    assert(state == 'pendding', 'Transaction state is not pendding');
+    if (state !== 'pendding') throw new PlugitError('Transaction state is not pendding');
     //Check operation registry;
-    assert(component[operation], `Operation [${operation}] of component [${component.name}] has not registed!`);
+    if (typeof component[operation] !== 'function') throw new PlugitError(`Operation [${operation}] of component [${component.name}] has not registed!`);
     //Bind transaction to component and generate an Action;
-    yield component._bindTransaction(this._id);
+    yield component.bindTransaction(this.id);
     const actionId = new ObjectId();
-    yield TransactionModel.findByIdAndUpdate(this._id, {
+    const componentInfo = yield component.info();
+    yield this.model.findByIdAndUpdate(this.id, {
       $push: {
         actions: {
           _id: actionId,
           component: component.name,
           instance: component.id,
           operation,
-          prev:  yield component.info()
+          prev: componentInfo
         }
       }
     });
     //Execute the action.
-    const componentInfo = yield component.info();
-    assert(!componentInfo || componentInfo.transaction == this._id, 'The transaction has no access to instance');
-    yield TransactionModel.findOneAndUpdate({ _id: this._id, 'actions._id': actionId }, { $set: { 'actions.$.state': 'applied' } });
+    if (componentInfo && componentInfo.transaction != this.id) throw new PlugitError('The transaction has no access to instance');
     yield component[operation](...params);
+    yield this.model.findOneAndUpdate({ _id: this.id, 'actions._id': actionId }, { $set: { 'actions.$.state': 'applied' } });
   }
 
   * _apply() {
     const {state, actions} = yield this.info();
-    assert(state == 'pendding', 'Transaction state is not pendding');
-    assert(actions.every(action => action.state == 'applied'), 'Some actions has not applied');
-    yield TransactionModel.findByIdAndUpdate(this._id, { $set: { state: 'applied' } });
+    if (state !== 'pendding') throw new PlugitError('Transaction state is not pendding');
+    if (actions.some(action => action.state !== 'applied')) throw new PlugitError('Some actions has not applied');
+    yield this.model.findByIdAndUpdate(this.id, { $set: { state: 'applied' } });
     yield this._commitAllActions();
   }
 
   * _commitAllActions() {
     const {state, actions} = yield this.info();
-    assert(state == 'applied', 'Transaction state is not applied');
+    if (state !== 'applied') throw new PlugitError('Transaction state is not applied');
     for (let action of actions) {
-      yield TransactionModel.findOneAndUpdate({ _id: this._id, 'actions._id': action._id }, { $set: { 'actions.$.state': 'committed' } });
+      yield this.model.findOneAndUpdate({ _id: this.id, 'actions._id': action._id }, { $set: { 'actions.$.state': 'committed' } });
     }
   }
 
@@ -81,45 +100,56 @@ class Transaction {
       yield this._apply();
     }
     const {actions, state} = yield this.info();
-    assert(state == 'applied', 'Transaction state is not applied');
-    assert(actions.every(action => action.state == 'committed'), 'Some actions has not committed');
+    if (state !== 'applied') throw new PlugitError('Transaction state is not applied');
+    if (actions.some(action => action.state !== 'committed')) throw new PlugitError('Some actions has not committed');
     yield this._unbindAllInstances();
-    yield TransactionModel.findByIdAndUpdate(this._id, { $set: { state: 'committed' } });
+    yield this.model.findByIdAndUpdate(this.id, { $set: { state: 'committed' } });
+  }
+
+  _attachComponent(action) {
+    const Component = this.components[action.component];
+    if (!Component) throw new PlugitError(`Component [${action.component}] is not defined!`);
+    const component = new Component();
+    component.id = action.instance;
+    //Bind model to component by modelName set in Component;
+    if (['ComponentMap', 'ComponentRegistry', 'PluginMap', 'PluginRegistry', 'Transaction'].includes(component.modelName)) throw new PlugitError(`Component model [${component.modelName}] is an iternal component, do not use it in custom component!`);
+    const model = this.models[component.modelName];
+    if (!(model.base instanceof mongoose.constructor)) throw new PlugitError('model must be an instance of mongoose model');
+    component.model = model;
+    return component;
   }
 
   * _cancelAllActions() {
     const {actions, state} = yield this.info();
-    assert(state == 'rollback', 'Transaction state is not rollback');
+    if (state !== 'rollback') throw new PlugitError('Transaction state is not rollback');
     for (let action of actions.reverse()) {
       if (['applied', 'committed'].includes(action.state)) {
-        const Component = global.components[action.component];
-        assert(Component, `Component ${action.component} is not defined!`);
-        const component = new Component();
-        component.id = action.instance;
+        const component = this._attachComponent(action);
         const componentInfo = yield component.info();
-        assert(!componentInfo || componentInfo.transaction == this._id, 'The transaction has no access to instance');
-        const model = component.model;
-        if (action.prev) {
-          yield model.findByIdAndUpdate(component.id, action.prev, { upsert: true, overwrite: true });
+        if (componentInfo && componentInfo.transaction !== this.id) throw new PlugitError('The transaction has no access to instance');
+        if (component.rollback) {
+          yield component.rollback(action.prev);
         } else {
-          yield model.findByIdAndRemove(component.id);
+          const model = component.model;
+          if (action.prev) {
+            yield model.findByIdAndUpdate(component.id, action.prev, { upsert: true, overwrite: true });
+          } else {
+            yield model.findByIdAndRemove(component.id);
+          }
         }
       }
-      yield TransactionModel.findOneAndUpdate({ _id: this._id, 'actions._id': action._id }, { $set: { 'actions.$.state': 'cancelled' } });
+      yield this.model.findOneAndUpdate({ _id: this.id, 'actions._id': action._id }, { $set: { 'actions.$.state': 'cancelled' } });
     }
   }
 
   * _unbindAllInstances() {
     const {actions, state} = yield this.info();
-    assert(['applied', 'rollback'].includes(state), 'Transaction state is not rollback or applied');
+    if (!['applied', 'rollback'].includes(state)) throw new PlugitError('Transaction state is not rollback or applied');
     for (let action of actions) {
-      const Component = global.components[action.component];
-      assert(Component, `Component [${action.component}] is not defined!`);
-      const component = new Component();
-      component.id = action.instance;
+      const component = this._attachComponent(action);
       const componentInfo = yield component.info();
-      if (!componentInfo || componentInfo.transaction == this._id) {
-        yield component._unbindTransaction(this._id);
+      if (!componentInfo || componentInfo.transaction == this.id) {
+        yield component.unbindTransaction(this.id);
       }
     }
   }
@@ -128,10 +158,10 @@ class Transaction {
     const {state} = yield this.info();
     switch (state) {
       case 'init':
-        yield TransactionModel.findByIdAndUpdate(this._id, { $set: { state: 'rollback' } });
+        yield this.model.findByIdAndUpdate(this.id, { $set: { state: 'rollback' } });
         break;
       case 'pendding':
-        yield TransactionModel.findByIdAndUpdate(this._id, { $set: { state: 'rollback' } });
+        yield this.model.findByIdAndUpdate(this.id, { $set: { state: 'rollback' } });
         yield this._cancelAllActions();
         break;
       case 'rollback':
@@ -147,44 +177,50 @@ class Transaction {
   * cancel() {
     yield this._rollback();
     const {actions, state} = yield this.info();
-    assert(state == 'rollback', 'Transaction state is not rollback');
-    assert(actions.every(action => action.state == 'cancelled'), 'Some actions has not cancelled');
+    if (state !== 'rollback') throw new PlugitError('Transaction state is not rollback');
+    if (actions.some(action => action.state !== 'cancelled')) throw new PlugitError('Some actions has not cancelled');
     yield this._unbindAllInstances();
-    yield TransactionModel.findByIdAndUpdate(this._id, { $set: { state: 'cancelled' } });
+    yield this.model.findByIdAndUpdate(this.id, { $set: { state: 'cancelled' } });
   }
 
   // It's an unsafe operation!!! Some versions of data will lost! 
   * revert() {
     const {actions, state} = yield this.info();
-    assert(state == 'committed', 'Transaction state is not committed');
+    if (state !== 'committed') throw new PlugitError('Transaction state is not committed');
     for (let action of actions.reverse()) {
-      const Component = global.components[action.component];
-      assert(Component, `Component ${action.component} is not defined!`);
-      const component = new Component();
-      component.id = action.instance;
-      const componentInfo = yield component.info();
-      assert(!componentInfo || !componentInfo.transaction || componentInfo.transaction == this._id, 'The transaction has no access to instance');
-      const model = component.model;
-      if (action.prev) {
-        yield model.findByIdAndUpdate(component.id, action.prev, { upsert: true, overwrite: true });
-      } else {
-        yield model.findByIdAndRemove(component.id);
+      if (['applied', 'committed'].includes(action.state)) {
+        const component = this._attachComponent(action);
+        const componentInfo = yield component.info();
+        if (componentInfo && componentInfo.transaction !== this.id) throw new PlugitError('The transaction has no access to instance');
+        if (component.rollback) {
+          yield component.rollback(action.prev);
+        } else {
+          const model = component.model;
+          if (action.prev) {
+            yield model.findByIdAndUpdate(component.id, action.prev, { upsert: true, overwrite: true });
+          } else {
+            yield model.findByIdAndRemove(component.id);
+          }
+        }
       }
-      yield TransactionModel.findOneAndUpdate({ _id: this._id, 'actions._id': action._id }, { $set: { 'actions.$.state': 'reverted' } });
+      yield this.model.findOneAndUpdate({ _id: this.id, 'actions._id': action._id }, { $set: { 'actions.$.state': 'reverted' } });
     }
-    yield TransactionModel.findByIdAndUpdate(this._id, { $set: { state: 'reverted' } });
+    yield this.model.findByIdAndUpdate(this.id, { $set: { state: 'reverted' } });
   }
 }
 
 Transaction.create = function* (id) {
-  let t = new Transaction(id);
+  const t = new Transaction(id);
   yield t._create();
   return t;
 };
 
 Transaction.middleware = {
   inject: function* (next) {
-    this.transaction = yield Transaction.create();
+    if (!this.plugit) throw new PlugitError('Please inject plugit into koa context first');
+    const components = this.plugit.components;
+    const models = this.plugit.models;
+    this.transaction = yield new Transaction(models, components).create();
     yield this.transaction.pend();
     try {
       yield next;
